updated getAllPosts

+++ lib/classes/POI.dart
@@ -4,15 +4,15 @@ import 'package:softshares/classes/publication.dart';
 import 'package:softshares/classes/user.dart';
 
 class POI extends Publication {
-  int _aval;
+  double _aval;
 
   POI(User user, User? admin, String desc, String title, bool validated,
       int subCategory, DateTime postDate,  File? imgPath, String? location, this._aval,)
       : super(user, admin, desc, title, validated, subCategory, postDate, imgPath, location);
 
-  int get aval => _aval;
+  double get aval => _aval;
 
-  set aval(int value) {
+  set aval(double value) {
     _aval = value;
   }
 }
